# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Roblox Vector Search API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/commands/gather-games": {
      "post": {
        "summary": "Start a job to gather games from Roblox API",
        "operationId": "gatherGames",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Job started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "running",
                            "completed",
                            "failed"
                          ]
                        }
                      },
                      "required": [
                        "jobId",
                        "message",
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "success": true,
                  "data": {
                    "jobId": "string",
                    "message": "string",
                    "status": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Global error response",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Error message"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "data": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/generateEmbeddings": {
      "post": {
        "summary": "Start a job to generate vector embeddings for all games",
        "operationId": "generateEmbeddings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Job started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "jobId",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "success": true,
                  "data": {
                    "jobId": "string",
                    "message": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Global error response",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Error message"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "data": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/similar-games/{universeId}": {
      "get": {
        "summary": "Find games similar to a given game based on embedding similarity.",
        "operationId": "getSimilarGames",
        "parameters": [
          {
            "name": "universeId",
            "in": "path",
            "description": "Universe ID of the game to find similar games for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of similar games to return (default: 10)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of similar games sorted by similarity score (highest first)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "universeId": {
                            "type": "number"
                          },
                          "rootPlaceId": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "gameplayDescription": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "similarity": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "universeId",
                          "rootPlaceId",
                          "name",
                          "description",
                          "gameplayDescription",
                          "similarity"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "universeId": 0,
                      "rootPlaceId": 0,
                      "name": "string",
                      "description": null,
                      "gameplayDescription": null,
                      "similarity": 0
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Game not found or no embeddings available",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Global error response",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Error message"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "data": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}